refactor(format): break image formatting into separate fn
