+++ baker/formatWordpressPost.tsx
@@ -1,13 +1,11 @@
 import * as cheerio from "cheerio"
 import urlSlug from "url-slug"
-import * as lodash from "lodash"
 import React from "react"
 import ReactDOMServer from "react-dom/server.js"
 import { BAKED_BASE_URL, HTTPS_ONLY } from "../settings/serverSettings.js"
 import { getTables } from "../db/wpdb.js"
 import Tablepress from "../site/Tablepress.js"
 import { GrapherExports } from "../baker/GrapherBakingUtils.js"
-import * as path from "path"
 import { RelatedCharts } from "../site/blocks/RelatedCharts.js"
 import {
     BLOCK_WRAPPER_DATATYPE,
@@ -27,6 +25,7 @@ import {
     DEEP_LINK_CLASS,
     extractDataValuesConfiguration,
     formatDataValue,
+    formatImages,
     parseKeyValueArgs,
 } from "./formatting.js"
 import { mathjax } from "mathjax-full/js/mathjax.js"
@@ -461,52 +460,7 @@ export const formatWordpressPost = async (
         }
     })
 
-    // Image processing
-    // Assumptions:
-    // - original images are not uploaded with a suffix "-[number]x[number]"
-    //   (without the quotes).
-    // - variants are being generated by wordpress when the original is uploaded
-    // - images are never legitimate direct descendants of <a> tags.
-    //   <a><img /></a> is considered deprecated (was used to create direct links to
-    //   the full resolution variant) and wrapping <a> will be removed to prevent
-    //   conflicts with lightboxes. Chosen over preventDefault() in front-end code
-    //   to avoid clicks before javascript executes.
-    for (const el of cheerioEl("img").toArray()) {
-        // Recreate source image path by removing automatically added image
-        // dimensions (e.g. remove 800x600).
-        const src = el.attribs["src"]
-        const parsedPath = path.parse(src)
-        let originalFilename = ""
-        if (parsedPath.ext !== ".svg") {
-            originalFilename = parsedPath.name.replace(/-\d+x\d+$/, "")
-            const originalSrc = path.format({
-                dir: parsedPath.dir,
-                name: originalFilename,
-                ext: parsedPath.ext,
-            })
-            el.attribs["data-high-res-src"] = originalSrc
-        } else {
-            originalFilename = parsedPath.name
-        }
-
-        // Remove wrapping <a> tag, conflicting with lightbox (cf. assumptions above)
-        if (el.parent.tagName === "a") {
-            const $a = cheerioEl(el.parent)
-            $a.replaceWith(cheerioEl(el))
-        }
-
-        // Add alt attribute
-        if (!el.attribs["alt"]) {
-            el.attribs["alt"] = lodash.capitalize(
-                originalFilename.replace(/[-_]/g, " ")
-            )
-        }
-
-        // Lazy load all images, unless they already have a "loading" attribute.
-        if (!el.attribs["loading"]) {
-            el.attribs["loading"] = "lazy"
-        }
-    }
+    formatImages(cheerioEl)
 
     // Table of contents and deep links
     const tocHeadings: TocHeading[] = []
